requirements.txt
pystray==0.19.4
Pillow==10.1.0
psutil==5.9.6
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.9
python-dotenv==1.0.0
paho-mqtt==1.6.1
pywin32==306; platform_system == "Windows" 